Player Rankings update.
Refactored player rankings to now use live data. First pings the ESPN API for our current scoring period.

We update our player data state to map through the values and list the player statistics.

If the current period has no data for the first player in the array that means that there currently are no active games and renders out the stats as "-".

I would like to refactor the initial league data state and move it up to the parent component as there are two API calls to get league data which is redundant and unnecessary.

This can be done in a single call and just pass down the data as props.