Major Update.
Player watchlist is now working properly. Ports changed from 6000 to 6001 as 6000 is an "unsafe" port used by Chrome. Refactor allows for the watchlist to render properly.

Introduced a major change in component rendering. Watchlist component is now the parent component to our player rankings. This was necessary in order to reduce the number of component re-renders. Prior to this change, watchlist did not re-render as the player rankings component held the add column. In order for us to be able to re-render the component without the use of useContext, it's necessary to state to be involved. It's not possible for us to manipulate states when the components live under the same parent component, thus we allowed for our player rankings to be the child component to our watchlist.

Doing this required a re-arrangement of the components and their styling so that it would still look the same upon render. I also move the addToWatchlist function from the player rankings up into the watchlist component. After doing this, I passed the function down as a prop to the player-rankings allowing for the functionality to stay consistent.

Then, I added a new state called watchlist after succesfully adding a player to our watchlist database and receiving the return value. I added in an update to the state to rerender the component which works successfully.

The question "but you still re-render the player rankings which isn't optimal" may come up, but in this case it works out well as an additional re-render of our player rankings is not a bad thing. Player Rankings works with live data and requires multiple re-renders to retrieve the most up to date information from the ESPN API. Thus, this would actually be efficient.

The next steps for this app would be to create an automatic component re-render of the player rankings every 30 seconds to grab the most up to date and accurate information without the need for the user to click refresh.