Updated Watchlist.
Watchlist updated to include only three columns. Initially I wanted the watchlist to be a overflow type view, however I realized that with an overflow it would have a fixed width and height which would not allow for responsive rendering.

Instead I opted to have a fixed card heights and lengths with overflow for the player names as they would definitely go beyond but also keep the responsive style of the actual component.

If the watchlist becomes too small bootstrap adds in an automatic vertical overflow which still works fine for my purposes.