Component Update.
My lovely UX Designer girlfriend advised me that a click "next" and "prev" functionality is cool but from a design perspective is very ugly.

She asked me to refactor the code to utilize a scrollable table instead with a fixed header. She wanted me to remove the scrollbar but it does not seem like it's feasible at least across all browsers so thus we kept it in as it doesn't look too bad.

Player watchlist is going to be refactored as well to include an overflow section to be scrollable via the x axis if possible. Will investigate and work on that next.

Player rankings css essentially complete. Responsive component still works well together with the refactored style.