Watchlist/Rankings Style Update + Functionality Update.
Watchlist now renders a row that says "Watchlist Empty!" When the database returns no players. Previously, it showed a render of four different empty images and player names that would say "Add Player." Visually this made no sense to keep as we aren't rendering out 4 players each time and its more clear to the user to add players as needed.

The Player Rankings have also been updated to now show a different favorites icon instead when the player is added to our watchlist. It shows a gold star to indicated that we have already once selected that option.