Player Rankings update.
Changed location of adding players to watchlist. Move it next to the player and edited the column title to "PWL" which stands for player watchlist.

Edited the images from a star to a "plus" mark and a "check" mark to indicated when a player is already added and when we want to add a player.

Watchlist is working well still with this and renders perfectly.

Working on removal section on watchlist next.