Overflow Update.
Noticed while on MacOS with a plugged in mouse vs a bluetooth mouse, the plugged in mouse renders the MacOS default setting to show overflow scroll columns. When not using auto it would render out the scroll column before as white boxes which do not look good.

Changed to overflow auto to manage hiding the setting until overflow is actually needed which is in very rare cases.