Watchlist Image Update.
Watchlist now renders player images. The render size works perfectly for the size of the app and the % make it able to render at multiple different sizes.

Image background is rounded to match our component and the background color matches the titles.

Final update is now to add the watchlist removal icon.